refactor(types): moved Sources to search and searchData results are Sources not the Annotation
