+++ lib/types/search.ts
@@ -1,5 +1,12 @@
-import type { SourcesAnnotation } from './annotations';
 import type { ToolInvocation, ToolStatus } from './tool';
+
+export interface SearchToolInvocation extends ToolInvocation {
+  toolName: 'search';
+  args: {
+    query: string;
+  };
+  result?: string;
+}
 
 export type SearchType = 'web' | 'database' | 'general';
 
@@ -8,13 +15,14 @@ export interface SearchData {
   query: string;
   type: SearchType;
   status: ToolStatus;
-  results?: SourcesAnnotation;
+  results?: Source[];
 }
 
-export interface SearchToolInvocation extends ToolInvocation {
-  toolName: 'search';
-  args: {
-    query: string;
-  };
-  result?: string;
+export interface Source {
+  type: string;
+  text: string;
+  headings?: string[];
+  url?: string;
+  file_reference?: string;
+  position?: number;
 }
